// web/src/app/model/form/update-form-model.ts
export class UpdateFormModel extends BaseModel {
    formType: FormType;
    formTypeId: number;
    gRS: GRS;
    gRSId: number;
    dPID: string;
    sampleQTY: number;
    dateCode: string;
    dateCodeActual: boolean;
    finalResult: boolean;
    remark: string;
    userName: string;
    qN: string;
    iPVersion: string;
    refDoc: string;
    qNSoftcopy: string;
    gRSNoType: string;
    batchDC: string;
    storageLoc: string;
    requireStatus: number;
    packQty: number;
    oddQtyPerBox: string;
    receivingInfo: string;
    dANo: string;
    packSampleQty: number;
    buyerID: number;
    divisionId: number;
    sAPDefectTypeID: number;
    vISDefectTypeID: number;
    fUNDefectTypeID: number;
    funMeasureTypeID: number;
    dIMDefectTypeID: number;
    mPositionDefectTypeID: number;
    lPositionDefectTypeID: number;
    dateCodeDefectTypeID: number;
    bowTwistDefectTypeID: number;
    funParaDefectTypeID: number;
    microDefectTypeID: number;
    testReportDefectTypeID: number;
    specDefectType: string;
    lastFormID: number;
    lockFlag: number;
    lockUser: string;
    dCRequirement: string;
constructor(form?: Form) {
    super(form);
    if (form) {
        this.formType = form.formType;
        this.formTypeId = form.formTypeId;
        this.gRS = form.grs;
        this.gRSId = form.grsId;
        this.dPID = form.dpid;
        this.sampleQTY = form.sampleQTY;
        this.dateCode = form.dateCode;
        this.dateCodeActual = form.dateCodeActual;
        this.finalResult = form.finalResult;
        this.remark = form.remark;
        this.userName = form.userName;
        this.qN = form.qn;
        this.iPVersion = form.ipVersion;
        this.refDoc = form.refDoc;
        this.qNSoftcopy = form.qnSoftcopy;
        this.gRSNoType = form.grsNoType;
        this.batchDC = form.batchDC;
        this.storageLoc = form.storageLoc;
        this.requireStatus = form.requireStatus;
        this.packQty = form.packQty;
        this.oddQtyPerBox = form.oddQtyPerBox;
        this.receivingInfo = form.receivingInfo;
        this.dANo = form.daNo;
        this.packSampleQty = form.packSampleQty;
        this.buyerID = form.buyerID;
        this.divisionId = form.divisionId;
        this.sAPDefectTypeID = form.sapDefectTypeID;
        this.vISDefectTypeID = form.vISDefectTypeID;
        this.fUNDefectTypeID = form.fUNDefectTypeID;
        this.funMeasureTypeID = form.funMeasureTypeID;
        this.dIMDefectTypeID = form.dIMDefectTypeID;
        this.mPositionDefectTypeID = form.mPositionDefectTypeID;
        this.lPositionDefectTypeID = form.lPositionDefectTypeID;
        this.dateCodeDefectTypeID = form.dateCodeDefectTypeID;
        this.bowTwistDefectTypeID = form.bowTwistDefectTypeID;
        this.funParaDefectTypeID = form.funParaDefectTypeID;
        this.microDefectTypeID = form.microDefectTypeID;
        this.testReportDefectTypeID = form.testReportDefectTypeID;
        this.specDefectType = form.specDefectType;
        this.lastFormID = form.lastFormID;
        this.lockFlag = form.lockFlag;
        this.lockUser = form.lockUser;
        this.dCRequirement = form.dcRequirement;
    } else {
        this.formType = null;
        this.formTypeId = 0;
        this.gRS = null;
        this.gRSId = 0;
        this.dPID = Constants.Empty;
        this.sampleQTY = 0;
        this.dateCode = Constants.Empty;
        this.dateCodeActual = false;
        this.finalResult = false;
        this.remark = Constants.Empty;
        this.userName = Constants.Empty;
        this.qN = Constants.Empty;
        this.iPVersion = Constants.Empty;
        this.refDoc = Constants.Empty;
        this.qNSoftcopy = Constants.Empty;
        this.gRSNoType = Constants.Empty;
        this.batchDC = Constants.Empty;
        this.storageLoc = Constants.Empty;
        this.requireStatus = 0;
        this.packQty = 0;
        this.oddQtyPerBox = Constants.Empty;
        this.receivingInfo = Constants.Empty;
        this.dANo = Constants.Empty;
        this.packSampleQty = 0;
        this.buyerID = 0;
        this.divisionId = 0;
        this.sAPDefectTypeID = 0;
        this.vISDefectTypeID = 0;
        this.fUNDefectTypeID = 0;
        this.funMeasureTypeID = 0;
        this.dIMDefectTypeID = 0;
        this.mPositionDefectTypeID = 0;
        this.lPositionDefectTypeID = 0;
        this.dateCodeDefectTypeID = 0;
        this.bowTwistDefectTypeID = 0;
        this.funParaDefectTypeID = 0;
        this.microDefectTypeID = 0;
        this.testReportDefectTypeID = 0;
        this.specDefectType = Constants.Empty;
        this.lastFormID = 0;
        this.lockFlag = 0;
        this.lockUser = Constants.Empty;
        this.dCRequirement = Constants.Empty;
    }
  }
}